Various fixes in SY feature page.
M2Eclipse -> Maven (since m2eclipse does not exist anymore and we use Maven on the other feature page)
schema -> schemas
Moved image:: just after === title so the layout is uniform for the image/titles.
Brought back image for top paragraph - it is repeated for the graphical editor but crates a balance in the layout.