JBIDE-18395 - Improve layout / Show CAT
Changed the content of /index.html:
- showing 3 hightlighted features on 1st row under the 'hero-section'
- showing the CAT on second row

"Latest News" and "Upcoming Events" sections are pushed on the right side with a grey background color.

Converted some topnav menu items into dropdowns for a quicker access to the specific downloads, forums, etc.
The intermediate pages such as /downloads still exist but are not a mandatory hop anymore.