JBIDE-23779 - Archive old JBoss Tools versions on Downloads and What's New pages
Reverted the products that were commented out: as there are some N&N for
these versions, the site generator extensions assume that the product
is going to be released soon, hence they all appear in a "coming soon"
state.
The current solution is to mark them all as "archived: true" to remove
them from the main pages. They still appear in the /downloads/overview
page, though.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>